requirements.txt
requests
google-cloud-storage
functions-framework==3.*
python-dotenv
pandas
google-cloud-bigquery
pytz
pyarrow